[Hard] Assignment #3 (#3)
You decide to make a simple mobile application to simulate the birds’ sounds you hear in the morning. After planning, you’ve decided on a single screen with three buttons. Each button will show the bird's name on the screen and print its sound to the console. Tapping a new button will replace the previous bird's name on the screen, and only the new bird’s sounds will print to the console.

Instructions

Create a single-screen app using coroutines in Compose. Add three buttons, each representing a different bird. Create a composable function for each bird that displays the bird’s name and launches a coroutine to print the bird's sound to the console. Only render the composable of the bird represented by the last button tapped and ensure only the selected bird's sounds are printed.