[Medium] Assignment #2
You’ve taken a couple of pictures of the birds you hear each morning, and you decide to write an application to select one of these pictures and then set the app’s background to the predominant color in the image.

Instructions

Take the existing project shared below with a single screen containing a button and the RotatingBox composable. Tapping the button should launch a picker that allows the user to select an image. Once an image is selected, use the function findDominantColor to get the primary color in the image. Change the app’s background to this new color.

Fix the program so that when executed, the rotating box runs smoothly after selecting an image.