[Easy] Assignment #1 (#6)
You’ve created a file documenting your thoughts on your bird friends and want to write a short application to upload it to a server for safekeeping. You write up a small demo project before integrating the API and notice a bug…

Instructions

Find and fix the bug in the code below to terminate the program when it has been canceled.