[Hard] Assignment #1 (#9)
With your newfound knowledge of coroutines, you decide to replicate an authentication prompt that you’ve seen on your favorite banking application. Upon testing your code, you’ve found a couple of bugs...

Instructions

Closely study the project linked below, which includes a button that, when tapped, will display an authentication prompt to a user. The prompt should be dismissed after a specific time has passed. Locate and fix the following bugs that you’ve discovered:

The app crashes when the user has no authentication method enrolled on the device.

The biometric prompt is not automatically cancelled after the timeout period.

The button does not launch the biometric prompt once it has been cancelled once.