[Easy] Assignment #1 (#11)
You want to build a small countdown application that starts a countdown and updates the UI every second. First, you want to construct a demo application to determine the logic behind the countdown timer.

Instructions

Create a demo application with a single screen and a button labeled “Start Countdown.” When tapped, trigger a flow {} builder that emits the value ten, and then each second reduces that value and emits the new value until 0 is reached. The app should observe the emissions and display them on the screen.