[Easy] Assignment #1
You want to build a small countdown application that starts a countdown and updates the UI every second. First, you want to construct a demo application to determine the logic behind the countdown timer.

Instructions

Create a demo application with a single screen and a button labeled “Start Countdown.” When tapped, trigger a flow {} builder that emits the value ten, and then each second reduces that value and emits the new value until 0 is reached. The app should observe the emissions and display them on the screen.

[Medium] Assignment #2

Happy with the current implementation of your demo application, you now want to expand it and support configuration changes.

Instructions

Expand your previous application with a new UI component that captures a numeric value as input and uses that value as the starting point of your countdown timer. The countdown timer should not be affected by a configuration change (i.e., device rotation).