[Medium] Assignment #1 (#13)
You’ve received a new feature request that requires you to show a persistent banner in the app when the user loses internet connectivity. The banner should be automatically dismissed when the user regains internet connectivity.

Instructions

Use Android’s ConnectivityManager class to construct a flow that emits a Boolean representing whether the user has access to the Internet. Observe this state and indicate in the UI whether the user has or does not have access to the Internet.