[Medium] Assignment #2
Happy with the current implementation of your demo application, you now want to expand it and support configuration changes.

Instructions

Expand your previous application with a new UI component that captures a numeric value as input and uses that value as the starting point of your countdown timer. The countdown timer should not be affected by a configuration change (i.e., device rotation).