[Hard] Assignment #2
Instructions

Based on the provided code sample, use your knowledge you’ve gained in this course to spot the things that would be very valuable to test for you to have confidence that this code works correctly. You may or may not need to adjust the code to work for the test case.